Extract AppearanceTabContent for reuse in dialog hosts
