+++ frontend/src/components/ui/Settings/UserPreferencesDialog.tsx
@@ -4,7 +4,6 @@ import clsx from "clsx";
 import { useEffect, useId, useMemo, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 
-import { useTheme, type ThemeMode } from "@/components/providers/ThemeProvider";
 import { useAudioInputDevicePreference } from "@/hooks/audio/useAudioInputDevicePreference";
 import {
   fetchCompleteMcpServerOauth,
@@ -26,20 +25,19 @@ import { Alert } from "../Feedback/Alert";
 import { FormField, Input, Textarea } from "../Input";
 import { ModalBase } from "../Modal/ModalBase";
 import {
-  ComputerIcon,
   ErrorIcon,
   LockIcon,
   LinkIcon,
   LinkSlashIcon,
   MediaImageIcon,
   MenuScaleIcon,
-  MoonIcon,
   ResolvedIcon,
-  SunIcon,
   CheckCircleIcon,
   VoiceIcon,
   WarningCircleIcon,
 } from "../icons";
+
+import { AppearanceTabContent } from "./AppearanceTabContent";
 
 import type {
   McpServerStatus,
@@ -55,13 +53,6 @@ type PreferencesTab =
   | "audio"
   | "mcpServers"
   | "data";
-
-interface AppearanceOption {
-  description: string;
-  icon: ReactNode;
-  label: string;
-  value: ThemeMode;
-}
 
 interface UserPreferencesDialogProps {
   isOpen: boolean;
@@ -90,7 +81,6 @@ export function UserPreferencesDialog({
   const { enabled: personalizationEnabled, mcpServersTabEnabled } =
     useUserPreferencesFeature();
   const { enabled: audioTranscriptionEnabled } = useAudioTranscriptionFeature();
-  const { effectiveTheme, setThemeMode, themeMode } = useTheme();
   const {
     audioInputDeviceError,
     audioInputDevices,
@@ -319,45 +309,6 @@ export function UserPreferencesDialog({
     ),
     data: <LockIcon className="size-4" />,
   } satisfies Record<PreferencesTab, ReactNode>;
-
-  const appearanceOptions = [
-    {
-      value: "light",
-      label: t({
-        id: "preferences.dialog.appearance.theme.light.label",
-        message: "Light mode",
-      }),
-      description: t({
-        id: "preferences.dialog.appearance.theme.light.description",
-        message: "Always use the light theme.",
-      }),
-      icon: <SunIcon className="size-5" />,
-    },
-    {
-      value: "dark",
-      label: t({
-        id: "preferences.dialog.appearance.theme.dark.label",
-        message: "Dark mode",
-      }),
-      description: t({
-        id: "preferences.dialog.appearance.theme.dark.description",
-        message: "Always use the dark theme.",
-      }),
-      icon: <MoonIcon className="size-5" />,
-    },
-    {
-      value: "system",
-      label: t({
-        id: "preferences.dialog.appearance.theme.system.label",
-        message: "System theme",
-      }),
-      description: t({
-        id: "preferences.dialog.appearance.theme.system.description",
-        message: "Match your device appearance settings.",
-      }),
-      icon: <ComputerIcon className="size-5" />,
-    },
-  ] satisfies AppearanceOption[];
 
   /* eslint-disable lingui/no-unlocalized-strings -- Internal DOM ids, not user-facing copy */
   const tabIds = {
@@ -702,69 +653,7 @@ export function UserPreferencesDialog({
                 </p>
               </div>
 
-              <div
-                role="radiogroup"
-                aria-label={t({
-                  id: "preferences.dialog.appearance.theme.heading",
-                  message: "Color mode",
-                })}
-                className="grid gap-3"
-              >
-                {appearanceOptions.map((option) => {
-                  const isSelected = themeMode === option.value;
-
-                  return (
-                    <button
-                      key={option.value}
-                      type="button"
-                      role="radio"
-                      aria-checked={isSelected}
-                      className={clsx(
-                        "flex items-start gap-3 rounded-lg border p-4 text-left",
-                        "theme-transition focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-theme-focus",
-                        isSelected
-                          ? "border-theme-border-focus bg-theme-bg-hover text-theme-fg-primary"
-                          : "border-theme-border bg-theme-bg-primary text-theme-fg-secondary hover:bg-theme-bg-hover",
-                      )}
-                      onClick={() => setThemeMode(option.value)}
-                    >
-                      <span
-                        aria-hidden="true"
-                        className={clsx(
-                          "mt-0.5 flex size-9 shrink-0 items-center justify-center rounded-md border",
-                          isSelected
-                            ? "border-theme-border-focus bg-theme-bg-secondary text-theme-fg-primary"
-                            : "border-theme-border bg-theme-bg-secondary text-theme-fg-secondary",
-                        )}
-                      >
-                        {option.icon}
-                      </span>
-
-                      <span className="min-w-0 flex-1">
-                        <span className="flex items-center justify-between gap-3">
-                          <span className="font-medium">{option.label}</span>
-                          {option.value === "system" && isSelected ? (
-                            <span className="text-xs text-theme-fg-muted">
-                              {effectiveTheme === "dark"
-                                ? t({
-                                    id: "preferences.dialog.appearance.theme.current.dark",
-                                    message: "Currently dark",
-                                  })
-                                : t({
-                                    id: "preferences.dialog.appearance.theme.current.light",
-                                    message: "Currently light",
-                                  })}
-                            </span>
-                          ) : null}
-                        </span>
-                        <span className="mt-1 block text-sm text-theme-fg-muted">
-                          {option.description}
-                        </span>
-                      </span>
-                    </button>
-                  );
-                })}
-              </div>
+              <AppearanceTabContent />
             </section>
 
             {audioTranscriptionEnabled ? (
